requirements.txt
Django==3.0.3
django-webpack-loader==0.7.0
djangorestframework==3.11.0
pytz==2019.3